# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Dexa Home Test Backend API (Gateway)
  version: 1.0.0
  description: |
    Unified API specification for all backend endpoints exposed through the API Gateway.
    Base URL points to the gateway root, and all paths below are client-facing paths.

servers:
  - url: http://localhost:3000
    description: Local API Gateway

tags:
  - name: Auth
  - name: Employee
  - name: Attendance
  - name: Notification
    description: |
      Notification delivery is primarily via WebSocket (Socket.IO namespace `/admin`),
      not REST endpoints.

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Requires `x-api-key` header.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/refresh:
    get:
      tags: [Auth]
      summary: Refresh access token
      description: 'Send refresh token in `Authorization: Bearer <refresh_token>`.'
      security:
        - BearerAuth: []
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/logout:
    delete:
      tags: [Auth]
      summary: Logout current session
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employee/employees/me:
    get:
      tags: [Employee]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current employee profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employee/employees/me/profile:
    patch:
      tags: [Employee]
      summary: Update own profile (phone/photo)
      description: |
        Use `multipart/form-data` to upload photo.
        `photo` is binary; `phoneNumber` is optional.
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phoneNumber:
                  type: string
                  maxLength: 50
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employee/employees/me/password:
    patch:
      tags: [Employee]
      summary: Change own password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeMyPasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employee/employees/admin:
    get:
      tags: [Employee]
      summary: List employees (admin only)
      description: Requires user role `ADMIN_HR`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - in: query
          name: role
          schema:
            type: string
            enum: [EMPLOYEE, ADMIN_HR]
          required: false
        - in: query
          name: search
          schema:
            type: string
            maxLength: 200
          required: false
      responses:
        '200':
          description: Employee list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Employee]
      summary: Create employee (admin only)
      description: Requires user role `ADMIN_HR`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employee/employees/admin/{id}:
    get:
      tags: [Employee]
      summary: Get employee detail by ID (admin only)
      description: Requires user role `ADMIN_HR`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
      responses:
        '200':
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Employee]
      summary: Update employee (admin only)
      description: Requires user role `ADMIN_HR`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /attendance/attendances:
    post:
      tags: [Attendance]
      summary: Create attendance (IN/OUT)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceRequest'
      responses:
        '201':
          description: Attendance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /attendance/attendances/me:
    get:
      tags: [Attendance]
      summary: Get my attendance summary (grouped by date)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/StartDateQuery'
        - $ref: '#/components/parameters/EndDateQuery'
      responses:
        '200':
          description: Grouped attendance summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAttendanceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /attendance/attendances/me/today:
    get:
      tags: [Attendance]
      summary: Get my attendance status for today
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Today's attendance status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyTodayAttendanceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /attendance/attendances/admin:
    get:
      tags: [Attendance]
      summary: List all attendance summary (admin only)
      description: Requires user role `ADMIN_HR`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/StartDateQuery'
        - $ref: '#/components/parameters/EndDateQuery'
        - in: query
          name: employeeId
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Admin attendance summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAttendanceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  parameters:
    EmployeeIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
    PageQuery:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    StartDateQuery:
      in: query
      name: startDate
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-03-01'
    EndDateQuery:
      in: query
      name: endDate
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-03-19'

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Must include upper, lower, number, and symbol.

    LoginResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    RefreshResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string

    ChangeMyPasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          description: Must include upper, lower, number, and symbol.

    CreateEmployeeRequest:
      type: object
      required: [name, email, password, position]
      properties:
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 150
        password:
          type: string
          minLength: 8
        position:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [EMPLOYEE, ADMIN_HR]
        phoneNumber:
          type: string
          maxLength: 50

    UpdateEmployeeRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        position:
          type: string
          maxLength: 100
        phoneNumber:
          type: string
          maxLength: 50

    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        position:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, ADMIN_HR]
        phoneNumber:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    CreateAttendanceRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [IN, OUT]

    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        attendanceDate:
          type: string
          format: date-time
        attendanceTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [IN, OUT]
        createdAt:
          type: string
          format: date-time

    MyAttendanceDay:
      type: object
      properties:
        date:
          type: string
          example: '2026-03-19'
        hasCheckedIn:
          type: boolean
        hasCheckedOut:
          type: boolean
        firstInTime:
          type: string
          nullable: true
          format: date-time
        lastOutTime:
          type: string
          nullable: true
          format: date-time
        lastStatus:
          type: string
          nullable: true
          enum: [IN, OUT]
        totalWorkHours:
          type: number
          format: float

    MyAttendanceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MyAttendanceDay'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    MyTodayAttendanceResponse:
      type: object
      properties:
        date:
          type: string
          example: '2026-03-19'
        hasCheckedIn:
          type: boolean
        hasCheckedOut:
          type: boolean
        firstInTime:
          type: string
          nullable: true
          format: date-time
        lastOutTime:
          type: string
          nullable: true
          format: date-time
        lastStatus:
          type: string
          nullable: true
          enum: [IN, OUT]

    AdminAttendanceRow:
      type: object
      properties:
        date:
          type: string
        employeeId:
          type: string
        employeeName:
          type: string
          nullable: true
        hasCheckedIn:
          type: boolean
        hasCheckedOut:
          type: boolean
        firstInTime:
          type: string
          nullable: true
          format: date-time
        lastOutTime:
          type: string
          nullable: true
          format: date-time
        lastStatus:
          type: string
          nullable: true
          enum: [IN, OUT]
        totalWorkHours:
          type: number
          format: float

    AdminAttendanceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminAttendanceRow'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string